update navbar links
