+++ client/App.js
@@ -1,6 +1,8 @@
 import React from "react";
 import { createBrowserRouter, RouterProvider } from "react-router-dom"
 import Main from './pages/Main'
+import Saved from './pages/Saved'
+import Favorites from './pages/Favorites'
 import Root from './pages/Root'
 import "slick-carousel/slick/slick.css"
 import "slick-carousel/slick/slick-theme.css"
@@ -16,7 +18,10 @@ const router = createBrowserRouter([
       { path: "/", element: <Home /> },
       { path: "/main", element: <Main />},
       { path: "/login", element: <Login />},
-      { path: "/signup", element: <Signup />}
+      { path: "/signup", element: <Signup />},
+      { path: "/home", element: <Home />},
+      { path: "/favorites", element: <Favorites />},
+      { path: "/saved", element: <Saved />}
     ],
   },
 ]);
