+++ client/pages/Favorites.js
@@ -1,0 +1,10 @@
+export default function Saved() {
+
+  return (
+    <div className="Favorites">
+        <h1>TRENDY</h1>
+      <h3>your FAVORITES data ... </h3>
+      under construction
+    </div>
+);
+}
