+++ client/pages/Saved.js
@@ -1,0 +1,10 @@
+export default function Saved() {
+
+  return (
+    <div className="Saved">
+        <h1>TRENDY</h1>
+      <h3>your SAVED data ... </h3>
+      under construction
+    </div>
+);
+}
